refactor: 도메인 서비스 구현
