+++ src/main/java/kitchenpos/eatinorders/tobe/domain/order/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.tobe.domain;
+package kitchenpos.eatinorders.tobe.domain.order;
 
 public enum OrderStatus {
     WAITING, ACCEPTED, SERVED, COMPLETED;
