+++ src/main/java/kitchenpos/eatinorders/tobe/domain/order/OrderTableId.java
@@ -1,0 +1,12 @@
+package kitchenpos.eatinorders.tobe.domain.order;
+
+import java.util.UUID;
+
+public class OrderTableId {
+
+    private final UUID id;
+
+    public OrderTableId(UUID id) {
+        this.id = id;
+    }
+}
