+++ src/main/java/kitchenpos/eatinorders/tobe/domain/ordertable/NumberOfGuests.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.tobe.domain;
+package kitchenpos.eatinorders.tobe.domain.ordertable;
 
 import java.util.Objects;
