+++ src/test/java/kitchenpos/Fixtures2.java
@@ -1,6 +1,6 @@
 package kitchenpos;
 
-import kitchenpos.eatinorders.tobe.domain.OrderLineItem;
+import kitchenpos.eatinorders.tobe.domain.order.OrderLineItem;
 import kitchenpos.menus.tobe.domain.Menu;
 import kitchenpos.menus.tobe.domain.MenuGroup;
 import kitchenpos.menus.tobe.domain.MenuProduct;
